% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_storm.R
\name{energy_storm}
\alias{energy_storm}
\title{The total energy of a storm}
\usage{
energy_storm(ID, dat)
}
\arguments{
\item{ID}{a string of the ID of a specific storm to calculate the energy for}

\item{dat}{the dataframe we use}
}
\value{
a numeric value representing the total energy of the storm (in kilojoules)
}
\description{
Given an ID representing a specific storm and the dat saved in the package,
the function returns a numeric value meaning the
accumulated cyclone energy (in kilojoules) of that storm. The energy is
calculated as the sum of the maximum sustained wind speeds of the
storm at 6-hour intervals and then divided  by 10000 to make it more
manageable, according to the definition of accumulated cyclone energy
}
\examples{
data(dat)
energy_storm("AL032005", dat)

}
