% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfall_storm.R
\name{landfall_storm}
\alias{landfall_storm}
\title{Check if a storm made landfall in the United States}
\usage{
landfall_storm(ID)
}
\arguments{
\item{ID}{The ID of the storm}
}
\value{
A list containing a logical value indicating whether the storm made
landfall in the United States, and a data frame
with the interpolated storm track
and an additional column indicating
whether each coordinate is within the United States.
}
\description{
Given a storm ID, this function interpolates the storm track by using
the helper function \code{interpolate_storm_track}. By estimating the possible
storm track and longitude and latitude coordinates of every time point, the function
uses for loop to decide whether each point on the track has
landfall in the United States or not
}
\details{
The final result returns a list containing a Boolean indicator reporting landfall
and also a complete dataframe containing every interpolated point's information
including longitude, latitude coordinates, time, 34, 50, and 64 knot data
and the landfall indicator
}
\examples{
data(dat)
landfall_storm("AL182012")

}
